refactor popup files structure
