+++ src/popup/pages/Cabinet/Cabinet.ts
@@ -1,4 +1,4 @@
-import Toolbar from '../../../directives/Toolbar/Toolbar';
+import Toolbar from '../../directives/Toolbar/Toolbar';
 
 export default {
   template: require('./Cabinet.html'),
