+++ src/popup/pages/Cabinet/IpfsCabinet/SavedContent/SavedContent.ts
@@ -1,10 +1,3 @@
-import { AppWallet, CoinType, getIpfsHash, StorageVars } from '../../../../../../services/data';
-import { CyberD } from '../../../../../../services/cyberd';
-import EthData from '@galtproject/frontend-core/libs/EthData';
-
-const _ = require('lodash');
-const pIteration = require('p-iteration');
-
 export default {
   template: require('./SavedContent.html'),
   created() {
