+++ src/popup/pages/NewWallet/ChooseMethod/ChooseMethod.html
