+++ src/popup/pages/NewWallet/ChooseMethod/ChooseMethod.ts
