+++ src/popup/pages/NewWallet/CreateWallet/CreateWallet.html
