+++ src/popup/pages/NewWallet/NewWallet.html
