+++ src/popup/pages/NewWallet/Welcome/Welcome.html
