+++ src/popup/pages/NewWallet/Welcome/Welcome.ts
